integrate presentation layer with data layer part 1

+++ composeApp/src/commonMain/kotlin/data/network/response/task/step/Activity.kt
@@ -18,7 +18,7 @@ data class Activity(
     @SerialName("form")
     val form : Form,
     @SerialName("tag")
-    val tag : Int,
+    val tag : Long,
     @SerialName("form_id")
     val form_id : Int,
     )
